Clean up handle integration example
