+++ examples/handle_integration/main.rs
@@ -3,6 +3,7 @@ mod custom_asset;
 mod daemon;
 mod game;
 mod image;
+mod storage;
 
 fn main() {
     std::env::set_current_dir(PathBuf::from("examples/daemon_with_loader"))
